organize discord related routes + add txn to setDiscordBotConfig.ts
